Revert "Merge branch 'driver-fix' into 'main'"
This reverts merge request !1
